requirements.txt
chromadb
flask
InstructorEmbedding
langchain
langchain_experimental
llama-index
llama-index-embeddings-instructor
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
llama-index-vector-stores-qdrant
llama-index-indices-managed-vectara
llama-index-llms-together
pyyaml
unstructured[pdf]
unstructured-client
python-dotenv
sentence-transformers
cython
python-magic 
python-magic-bin
flask-cors